- Found me a bug
git-svn-id: svn://70.113.44.214/VanaMirror/branches/buis_vana_overhaul@3124 09c1253f-3d09-8e40-9719-4320e56601b5

+++ src/LoginServer/RankingCalculator.cpp
@@ -61,8 +61,8 @@ namespace RankingCalculator {
 	const string rankIfClause() {
 		std::ostringstream ret;
 		ret << "("
-			<< "	(job DIV 1000 <> 1 AND level <> " << Stats::PlayerLevels << ") "
-			<< "	OR (job DIV 1000 = 1 AND level <> " << Stats::CygnusLevels << ")"
+			<< "	(job DIV 1000 <> 1 AND level <> " << static_cast<int16_t>(Stats::PlayerLevels) << ") "
+			<< "	OR (job DIV 1000 = 1 AND level <> " << static_cast<int16_t>(Stats::CygnusLevels) << ")"
 			<< ") "
 			<< "AND @level = level AND @exp = exp, @rank, @real_rank + 1";
 		return ret.str();
